(v2.14.4) patch: client splits profile and portfolio forms
profile and portfolio where under `/profile` in a tab layout but this was a little messy because the nav bar also had a portfolio link pointing to profile.
to resolve this mess, i simply split the profile and portfolio so under the admin btn and removed the porfolio from the nav items.